+++ docs/changelog.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Place any unreleased changes here, that are subject to release in coming versions :).
 
+## 1.7.2 - 2025-06-12
+
+* feat: Added SAML support for Spring boot.
+
 ## 1.7.1 - 2025-06-11
 
 * feat: Added SMTP support for Spring boot.
