+++ examples/springboot/spring-boot_app/src/main/java/com/canonical/sampleapp/web/rest/SAMLController.java
@@ -1,3 +1,8 @@
+/*  
+ * Copyright 2025 Canonical Ltd.  
+ * See LICENSE file for licensing details.  
+ */
+
 package com.canonical.sampleapp.web.rest;
 
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
