--- examples/springboot/spring-boot_app/src/main/resources/metadata.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<md:EntityDescriptor xmlns:md="urn:oasis:names:tc:SAML:2.0:metadata" entityID="testentity">
-    <md:SPSSODescriptor protocolSupportEnumeration="urn:oasis:names:tc:SAML:2.0:protocol">
-        <md:KeyDescriptor use="signing">
-            <ds:KeyInfo xmlns:ds="http://www.w3.org/2000/09/xmldsig#">
-                <ds:X509Data>
-                    <ds:X509Certificate>MIIDazCCAlOgAwIBAgIUOh9/Y2UpCWDqNsbBWfPepa4R5NowDQYJKoZIhvcNAQELBQAwRTELMAkGA1UEBhMCQVUxEzARBgNVBAgMClNvbWUtU3RhdGUxITAfBgNVBAoMGEludGVybmV0IFdpZGdpdHMgUHR5IEx0ZDAeFw0yNTA1MDgwOTE5NDlaFw0yNjA1MDgwOTE5NDlaMEUxCzAJBgNVBAYTAkFVMRMwEQYDVQQIDApTb21lLVN0YXRlMSEwHwYDVQQKDBhJbnRlcm5ldCBXaWRnaXRzIFB0eSBMdGQwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDei33U7BOPHHVbwOO5T+88pBbYL3u3mVMhonen1JalFAF1sYrePWC7iCw+5gtD9A+x0/FSxUwwcnzgok4id1wxlQUwH6HRG1kmpadfhzmAlHLWz0lYoPx0coeLYJdcfRR8sqwpqVXdopnrh0/Edwvpcwpt5p2KZnLpP+hBSYzZ430fY3xWR4DOWw7N3ZPeQ0Qr+6TjJxUXfM/ZxOyosPxG84rb1Xt83CDQufarY5j/x0JqHA+7/AvsHJhi/LpGbvRrxH7kXNOn1rjeA0V/3Nbw+Llf61vvYSyGl8QDSP1Y3JzaZigK5GgT6jQXYjvlpuN/fPVyKu9xUSK2BN7ls73xAgMBAAGjUzBRMB0GA1UdDgQWBBTvumcxgFrOsxkvM9HqigMz06fIrzAfBgNVHSMEGDAWgBTvumcxgFrOsxkvM9HqigMz06fIrzAPBgNVHRMBAf8EBTADAQH/MA0GCSqGSIb3DQEBCwUAA4IBAQAyi1l240j+e/mVswC7Nupvs0avGKMUA9GuskpwkqPwaJDBZiT6zwfLJrhPyqfOde5Mxqqy+UaR8jKpHMXGyIKc/gUXNIkF5h7FhzuR+HdEk9i7dcUrmLaHE0Lqquh6q1IEhEwUgfPfyjw37AXUeOwlX1i9Q2jNKTC1W0zGFXGBjaz+uVELHu9q8PBwQ0d51wFgYLzITmyz5wSzKXavWnXHefKXaoem1vK3CSeqff0F1P6h0Lu+KydSgf5sTzXJUnwBhh8EaZ61RaKQ+T2JH07QQDbvqA0UpstISDgdKlMY0GeoJzFWsN40lJ2r17qoHKnnv3IL5X26x+li3MEGnDdn</ds:X509Certificate>
-                </ds:X509Data>
-            </ds:KeyInfo>
-        </md:KeyDescriptor>
-        <md:AssertionConsumerService Binding="urn:oasis:names:tc:SAML:2.0:bindings:HTTP-POST" Location="http://192.168.1.196:8081/login/saml2/sso/testentity" index="1"/>
-    </md:SPSSODescriptor>
-    <md:IDPSSODescriptor protocolSupportEnumeration="urn:oasis:names:tc:SAML:2.0:protocol">
-        <md:SingleSignOnService Binding="urn:oasis:names:tc:SAML:2.0:bindings:HTTP-Redirect" Location="http://192.168.1.196:8080/simplesaml/saml2/idp/SSOService.php"/>
-        <md:KeyDescriptor use="signing">
-            <ds:KeyInfo xmlns:ds="http://www.w3.org/2000/09/xmldsig#">
-                <ds:X509Data>
-                    <ds:X509Certificate>MIICmjCCAYICCQDX5sKPsYV3+jANBgkqhkiG9w0BAQsFADAPMQ0wCwYDVQQDDAR0ZXN0MB4XDTE5MTIyMzA5MDI1MVoXDTIwMDEyMjA5MDI1MVowDzENMAsGA1UEAwwEdGVzdDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAMdtDJ278DQTp84O5Nq5F8s5YOR34GFOGI2Swb/3pU7X7918lVljiKv7WVM65S59nJSyXV+fa15qoXLfsdRnq3yw0hTSTs2YDX+jl98kK3ksk3rROfYh1LIgByj4/4NeNpExgeB6rQk5Ay7YS+ARmMzEjXa0favHxu5BOdB2y6WvRQyjPS2lirT/PKWBZc04QZepsZ56+W7bd557tdedcYdY/nKI1qmSQClG2qgslzgqFOv1KCOw43a3mcK/TiiD8IXyLMJNC6OFW3xTL/BG6SOZ3dQ9rjQOBga+6GIaQsDjC4Xp7Kx+FkSvgaw0sJV8gt1mlZy+27Sza6d+hHD2pWECAwEAATANBgkqhkiG9w0BAQsFAAOCAQEAm2fk1+gd08FQxK7TL04O8EK1f0bzaGGUxWzlh98a3Dm8+OPhVQRi/KLsFHliLC86lsZQKunYdDB+qd0KUk2oqDG6tstG/htmRYD/S/jNmt8gyPAVi11dHUqW3IvQgJLwxZtoAv6PNs188hvT1WK3VWJ4YgFKYi5XQYnR5sv69Vsr91lYAxyrIlMKahjSW1jTD3ByRfAQghsSLk6fV0OyJHyhuF1TxOVBVf8XOdaqfmvD90JGIPGtfMLPUX4m35qaGAU48PwCL7L3cRHYs9wZWc0ifXZcBENLtHYCLi5txR8c5lyHB9d3AQHzKHMFNjLswn5HsckKg83RH7+eVqHqGw==</ds:X509Certificate>
-                </ds:X509Data>
-            </ds:KeyInfo>
-        </md:KeyDescriptor>
-    </md:IDPSSODescriptor>
-</md:EntityDescriptor>
