--- examples/springboot/spring-boot_app/src/main/resources/templates/saml.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="https://www.thymeleaf.org"
-    xmlns:sec="https://www.thymeleaf.org/thymeleaf-extras-springsecurity6">
-
-<head>
-    <title>Spring Boot and SAML</title>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-</head>
-
-<body>
-
-    <h1>Welcome</h1>
-    <p>You are successfully logged in as <span sec:authentication="name"></span></p>
-    <p>Your email address is <span th:text="${emailAddress}"></span>.</p>
-    <p>Your authorities are <span sec:authentication="authorities"></span>.</p>
-    <h2>All Your Attributes</h2>
-    <dl th:each="userAttribute : ${userAttributes}">
-        <dt th:text="${userAttribute.key}"></dt>
-        <dd th:text="${userAttribute.value}"></dd>
-    </dl>
-
-    <form th:action="@{/logout}" method="post">
-        <button id="logout" type="submit">Logout</button>
-    </form>
-
-</body>
-
-</html>
