+++ src/paas_charm/springboot/charm.py
@@ -176,9 +176,12 @@ def generate_saml_env(
     """
     if not relation_data:
         return {}
+
     return {
-        "spring.security.saml2.relyingparty.registration.testentity.assertingparty.metadata-uri": relation_data.metadata_url,
+        "spring.security.saml2.relyingparty.registration.testentity.assertingparty.metadata-uri": relation_data.metadata_url.unicode_string(),
         "spring.security.saml2.relyingparty.registration.testentity.entity-id": relation_data.entity_id,
+        "spring.security.saml2.relyingparty.registration.testentity.assertingparty.singlesignin.url": relation_data.single_sign_on_redirect_url,
+        "spring.security.saml2.relyingparty.registration.testentity.assertingparty.verification.credentials[0].certificate-location": "file:/app/saml.cert",
     }
 
 
@@ -269,6 +272,7 @@ class Charm(PaasCharm):
         base_dir = pathlib.Path("/app")
         state_dir = base_dir / "state"
         framework_config = typing.cast(SpringBootConfig, self.get_framework_config())
+
         return WorkloadConfig(
             framework=framework_name,
             container_name=WORKLOAD_CONTAINER_NAME,
@@ -290,6 +294,13 @@ class Charm(PaasCharm):
             A new App instance.
         """
         charm_state = self._create_charm_state()
+        if charm_state.integrations.saml:
+            if charm_state.integrations.saml.signing_certificate:
+                cert = charm_state.integrations.saml.signing_certificate
+                if not cert.startswith("-----BEGIN CERTIFICATE-----"):
+                    cert = f"-----BEGIN CERTIFICATE-----\n{cert}\n-----END CERTIFICATE-----"
+                self._container.push(self._workload_config.app_dir / "saml.cert", cert)
+
         return SpringBootApp(
             container=self._container,
             charm_state=charm_state,
