+++ tests/unit/springboot/test_integrations.py
@@ -51,3 +51,59 @@ def test_smtp_integration(
     assert environment["spring.mail.password"] is None
     assert environment["spring.mail.properties.mail.smtp.auth"] == "none"
     assert environment["spring.mail.properties.mail.smtp.starttls.enable"] == "false"
+
+
+def test_saml_integration(
+    base_state,
+) -> None:
+    """
+    arrange: add saml relation to the base state.
+    act: start the springboot charm and set springboot-app container to be ready.
+    assert: the springboot charm should have the saml related env variables.
+    """
+    base_state["relations"].append(
+        testing.Relation(
+            endpoint="saml",
+            interface="saml-integrator",
+            remote_app_data={
+                "entity_id": "http://example.com/entity",
+                "metadata_url": "http://example.com/metadata",
+                "x509certs": "cert1",
+                "single_sign_on_service_redirect_url": "http://example.com/sso",
+                "single_sign_on_service_redirect_binding": "urn:oasis:names:tc:SAML:2.0:bindings:HTTP-Redirect",
+            },
+        )
+    )
+    state = testing.State(**base_state)
+    context = testing.Context(
+        charm_type=SpringBootCharm,
+    )
+    out = context.run(context.on.config_changed(), state)
+    environment = list(out.containers)[0].plan.services["spring-boot"].environment
+    assert out.unit_status == testing.ActiveStatus()
+
+    saml_relation = out.get_relations("saml")
+    assert len(saml_relation) == 1
+
+    assert (
+        environment[
+            "spring.security.saml2.relyingparty.registration.testentity.assertingparty.metadata-uri"
+        ]
+        == "http://example.com/metadata"
+    )
+    assert (
+        environment["spring.security.saml2.relyingparty.registration.testentity.entity-id"]
+        == "http://example.com/entity"
+    )
+    assert (
+        environment[
+            "spring.security.saml2.relyingparty.registration.testentity.assertingparty.singlesignin.url"
+        ]
+        == "http://example.com/sso"
+    )
+    assert (
+        environment[
+            "spring.security.saml2.relyingparty.registration.testentity.assertingparty.verification.credentials[0].certificate-location"
+        ]
+        == "file:/app/saml.cert"
+    )
